feat: ES Module support
