+++ .eslintrc.json
@@ -6,7 +6,7 @@
   },
   "parser": "@typescript-eslint/parser",
   "parserOptions": {
-    "project": "tsconfig.json",
+    "project": ["tsconfig.json", "tsconfig.test.json"],
     "sourceType": "module"
   },
   "plugins": [
