+++ README.md
@@ -24,11 +24,18 @@ npm install @foo-software/lighthouse-persist
 
 Below are a few standard ways of using this module. See [parametes](#parameters) and [response payload](#response-payload) for more details.
 
+Note: to import in an ES Module enabled environment, you'll need to do something like this until we export an ES Module distribution:
+
+```javascript
+import lighthousePersistPackage from '@foo-software/lighthouse-persist';
+const { lighthousePersist } = lighthousePersistPackage;
+```
+
 ## Save Report to Local Directory.
 
 ```javascript
 const path = require('path');
-const lighthousePersist = require('@foo-software/lighthouse-persist').default;
+const { lighthousePersist } = require('@foo-software/lighthouse-persist');
 
 (async () => {
   const { localReport, result } = await lighthousePersist({
