+++ tsconfig.test.json
@@ -5,4 +5,6 @@
     "moduleResolution": "node",
     "target": "es2020"
   },
+  "exclude": ["node_modules"],
+  "include": ["**/*.ts", "**/*.test.ts"]
 }
